[REFACTOR]: S3Service 인터페이스로 분리
